requirements.txt
# =============================================================================
# Tegifa Legal — Runtime Dependencies
# =============================================================================

# --- Web Framework ---
streamlit
streamlit-authenticator
requests

# --- RAG Pipeline ---
llama-index
llama-index-vector-stores-chroma
llama-index-embeddings-huggingface
llama-index-llms-openai
llama-index-llms-ollama
chromadb
sentence-transformers
pypdf

# --- Document Processing ---
nougat-ocr
python-docx
torch
transformers

# --- Multi-Agent Architecture ---
langgraph
langchain
duckduckgo-search
ddgs

# --- Knowledge Graph ---
neo4j

# --- Database ---
sqlalchemy

# --- HTTP ---
pydantic
httpx
